+++ addContext.js
@@ -1,0 +1,6 @@
+/*
+ * Export addContext function
+ *
+ */
+
+module.exports = require('./dist/addContext');